# docs/openapi/openapi-baseline.json
{
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "description": "Standard error ``detail`` object for 4xx/5xx business and security failures.\n\nAttributes:\n    code: Project-wide stable error code.\n    key: Machine-readable key for clients.\n    message: Human-readable English message.\n    source: Origin layer (e.g. ``business``, ``security``).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "code",
          "key",
          "message",
          "source"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserCreateResponse": {
        "description": "User resource returned by create (201) and get (200); maps ORM :class:`~app.models.core.user.User`.\n\nAttributes:\n    client_uuid: Internal client identifier (UUID string).\n    created_at: Row creation timestamp (timezone-aware).\n    updated_at: Last update timestamp (timezone-aware).\n    is_row_invalid: Soft-invalidation flag (0 or 1).\n    invalidation_reason_uuid: FK to invalidation reason, if set.\n    system_user_id: External system user id when present.\n    system_uuid: Related system UUID when present.\n    username: Optional display username.\n    full_name: Required full name.\n    timezone: IANA timezone string stored for the user.",
        "properties": {
          "client_uuid": {
            "title": "Client Uuid",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "invalidation_reason_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Reason Uuid"
          },
          "is_row_invalid": {
            "title": "Is Row Invalid",
            "type": "integer"
          },
          "system_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System User Id"
          },
          "system_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Uuid"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "client_uuid",
          "created_at",
          "updated_at",
          "is_row_invalid",
          "invalidation_reason_uuid",
          "system_user_id",
          "system_uuid",
          "username",
          "full_name",
          "timezone"
        ],
        "title": "UserCreateResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Study App API",
    "title": "Study App API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/user/{system_user_id}": {
      "get": {
        "description": "Returns user profile by external `system_user_id`.\n\n### Example request (curl)\n```bash\ncurl -X GET 'http://127.0.0.1:8000/api/v1/user/2' \\\n  -H 'accept: application/json' \\\n  -H 'X-API-Key: ....'\n```",
        "operationId": "getUserBySystemUserId",
        "parameters": [
          {
            "in": "path",
            "name": "system_user_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 1,
              "title": "System User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "auth_required": {
                    "summary": "Auth header required",
                    "value": {
                      "code": "COMMON_401",
                      "key": "SECURITY_AUTH_REQUIRED",
                      "message": "Missing or invalid API key in header `X-API-Key`.",
                      "source": "security"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key header."
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "user_not_found": {
                    "summary": "No user with given system_user_id",
                    "value": {
                      "code": "USER_404",
                      "key": "USER_NOT_FOUND",
                      "message": "User with this `system_user_id` was not found.",
                      "source": "business"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "User not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "rate_limit_exceeded": {
                    "summary": "Too many requests",
                    "value": {
                      "code": "COMMON_429",
                      "key": "SECURITY_RATE_LIMIT_EXCEEDED",
                      "message": "Too many requests. Retry later.",
                      "source": "security"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Per-client request rate limit exceeded.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until request can be retried.",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Limit": {
                "description": "Rate-limit ceiling for current window.",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Remaining requests in current window.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get user by system_user_id",
        "tags": [
          "User"
        ]
      }
    },
...
  },
...
}